Improve the developer experience through better Invoke commands (#233)

+++ devutils/src/devutils/invoke_utils.py
@@ -5,6 +5,7 @@ import sys
 import tempfile
 import textwrap
 from contextlib import contextmanager
+from datetime import datetime
 from functools import lru_cache
 from itertools import chain
 from pathlib import Path
@@ -12,7 +13,7 @@ from typing import Dict, Iterator, List, Optional, Tuple
 
 from invoke import task
 
-ROOT = Path(__file__).parent.parent.parent
+ROOT = Path(__file__).absolute().parent.parent.parent
 REPOSITORY_URL = "https://github.com/robocorp/robocorp/tree/master/"
 
 
@@ -171,11 +172,14 @@ def build_common_tasks(
         args = " ".join(str(c) for c in cmd)
         return ctx.run(args, **options)
 
-    def poetry(ctx, *cmd):
+    def poetry(ctx, *cmd, verbose: bool = False):
         prefix = []
         if _use_conda():
             prefix.extend(["conda", "run", "--no-capture-output", "-n", CONDA_ENV_NAME])
         prefix.append("poetry")
+
+        if verbose:
+            prefix.append("-vv")
 
         return run(ctx, *prefix, *cmd)
 
@@ -197,15 +201,53 @@ def build_common_tasks(
                 )
 
     @task
-    def install(ctx):
-        """Install dependencies"""
+    def install(
+        ctx, local: Optional[str] = None, update: bool = False, verbose: bool = False
+    ):
+        """
+        Installs or updates dependencies.
+
+        Args:
+            local: A comma-separated list of local projects to install in develop mode.
+            update: Whether to update the dependencies.
+            verbose: Whether to run in verbose mode.
+        """
         _make_conda_env_if_needed()
-        poetry(ctx, "install")
+
+        if update:
+            poetry(ctx, "update", verbose=verbose)
+            return
+
+        projects = (
+            [package.replace("robocorp-", "") for package in local.split(",")]
+            if local
+            else None
+        )
+        if projects:
+            with mark_as_develop_mode(projects):
+                poetry(ctx, "lock --no-update")
+                poetry(ctx, "install", verbose=verbose)
+        else:
+            poetry(ctx, "install", verbose=verbose)
 
     @task
-    def devinstall(ctx):
-        """Install dependencies with deps in dev mode"""
+    def devinstall(ctx, verbose: bool = False):
+        """
+        Install the package in develop mode and its dependencies.
+
+        Args:
+            verbose: Whether to run in verbose mode.
+        """
         _make_conda_env_if_needed()
+
+        with mark_as_develop_mode(all_packages=True):
+            poetry(ctx, "lock --no-update")
+            poetry(ctx, "install", verbose=verbose)
+
+    @contextmanager
+    def mark_as_develop_mode(
+        projects: Optional[list[str]] = None, all_packages: bool = False
+    ):
         root_pyproject = root / "pyproject.toml"
         assert root_pyproject.exists(), f"Expected {root_pyproject} to exist."
 
@@ -217,21 +259,22 @@ def build_common_tasks(
             for pyproject in all_pyprojects:
                 roundtrip_py_project = RoundtripPyProject(pyproject)
                 roundtrips.append(roundtrip_py_project)
+
                 with roundtrip_py_project.update() as contents:
                     dependencies = contents["tool"]["poetry"]["dependencies"]
 
                     for key, value in tuple(dependencies.items()):
+                        if not key.startswith("robocorp-"):
+                            continue
+
                         # Changes something as:
                         # robocorp-log = "0.1.0"
                         # to:
-                        # robocorp-log = {path = "../log/", develop = true}
-                        if key.startswith("robocorp-"):
-                            name = key[len("robocorp-") :]
-                            value = dict(path=f"../{name}/", develop=True)
-                            dependencies[key] = value
-
-            poetry(ctx, "lock --no-update")
-            poetry(ctx, "install")
+                        # robocorp-log = {path = "../log/", develop = true
+                        name = key[len("robocorp-") :]
+                        if all_packages or (projects and name in projects):
+                            dependencies[key] = dict(path=f"../{name}/", develop=True)
+            yield
         finally:
             for roundtrip in roundtrips:
                 roundtrip.restore()
@@ -271,11 +314,16 @@ def build_common_tasks(
         poetry(ctx, f"run isort {targets}")
 
     @task
-    def test(ctx):
+    def test(ctx, test: Optional[str] = None):
         """Run unittests"""
         cmd = "run pytest -rfE -vv"
+
+        if test:
+            cmd += f" {test}"
+
         if parallel_tests:
             cmd += " -n auto"
+
         poetry(ctx, cmd)
 
     @task
@@ -318,26 +366,25 @@ def build_common_tasks(
             tmp.close()  # Fix for Windows
             poetry(ctx, f"run mypy --strict {tmp.name}")
 
-    @task(lint, typecheck, test)
-    def check_all(ctx):
-        """Run all checks"""
-        pass
-
     @task
     def build(ctx):
         """Build distributable .tar.gz and .wheel files"""
         poetry(ctx, "build")
 
     @task
-    def lock(ctx):
-        """Run poetry lock"""
-        poetry(ctx, "lock")
+    def publish(ctx, token: str):
+        """Publish to PyPI"""
+        poetry(ctx, f"config pypi-token.pypi {token}")
+        poetry(ctx, "publish")
 
     @task
-    def docs(ctx):
+    def docs(ctx, check: bool = False):
         """Build API documentation"""
         output_path = root / "docs" / "api"
-        output_path.mkdir(exist_ok=True)
+        if not output_path.exists():
+            print("Docs output path does not exist. Skipping...")
+            return
+
         for path in output_path.iterdir():
             path.unlink()
 
@@ -352,6 +399,30 @@ def build_common_tasks(
             f"--output-path {output_path}",
             package_name,
         )
+
+        if check:
+            assert not check_document_changes(ctx), "There are uncommitted docs changes"
+
+    def check_document_changes(ctx):
+        """
+        Check if there were new document changes generated by lazydocs and
+        are uncommited.
+
+        Returns:
+            True if there are new changes, False otherwise.
+        """
+        changed_files = (
+            run(ctx, "git --no-pager diff --name-only -- docs/api", hide=True)
+            .stdout.strip()
+            .splitlines()
+        )
+
+        return bool(changed_files)
+
+    @task(lint, typecheck, test)
+    def check_all(ctx):
+        """Run all checks"""
+        ctx.run("inv docs --check")
 
     @task
     def make_release(ctx):
@@ -390,7 +461,8 @@ def build_common_tasks(
             echo=True,
         )
 
-        print(f"Trigger the release with: git push origin {current_tag}")
+        print(f"Pushing tag: {current_tag}")
+        run(ctx, f"git push origin {current_tag}")
 
     @task
     def check_tag_version(ctx):
@@ -398,27 +470,56 @@ def build_common_tasks(
         Checks if the current tag matches the latest version (exits with 1 if it
         does not match and with 0 if it does match).
         """
-        import importlib
-
-        mod = importlib.import_module(package_name)
+        module_version = poetry(
+            ctx,
+            f"run python -c 'import {package_name}; print({package_name}.__version__)'",
+        ).stdout.strip()
 
         tag = get_tag(tag_prefix)
         version = tag[tag.rfind("-") + 1 :]
 
-        if mod.__version__ == version:
+        if module_version == version:
             sys.stderr.write(f"Version matches ({version}) (exit(0))\n")
             sys.exit(0)
         else:
             sys.stderr.write(
-                f"Version does not match ({tag_prefix}: {mod.__version__} != repo tag: {version}).\nTags:{get_all_tags(tag_prefix)}\n(exit(1))\n"
+                f"Version does not match ({tag_prefix}: {module_version} != repo tag: {version}).\nTags:{get_all_tags(tag_prefix)}\n(exit(1))\n"
             )
             sys.exit(1)
+
+    def update_changelog_file(file: Path, version: str):
+        """Update the changelog file with the new version and changes"""
+
+        with open(file, "r+") as stream:
+            content = stream.read()
+
+            new_version = f"## {version} - {datetime.today().strftime('%Y-%m-%d')}"
+            changelog_start = re.search(r"# Changelog", content).end()
+            unreleased_match = re.search(r"## Unreleased", content, flags=re.IGNORECASE)
+            double_newline = "\n\n"
+
+            new_content = content[:changelog_start] + double_newline + "## Unreleased"
+            if unreleased_match:
+                released_content = content.replace(
+                    unreleased_match.group(), new_version
+                )
+                new_content += released_content[changelog_start:]
+            else:
+                new_content += double_newline + new_version + content[changelog_start:]
+
+            stream.seek(0)
+            stream.write(new_content)
+            print("Changed: ", file)
 
     @task
     def set_version(ctx, version):
         """Sets a new version for the project in all the needed files"""
-        import re
-        from pathlib import Path
+        valid_version_pattern = re.compile(r"^\d+\.\d+\.\d+$")
+        if not valid_version_pattern.match(version):
+            print(
+                f"Invalid version: {version}. Must be in the format major.minor.hotfix"
+            )
+            return
 
         version_patterns = (
             re.compile(r"(version\s*=\s*)\"\d+\.\d+\.\d+"),
@@ -445,6 +546,8 @@ def build_common_tasks(
         # Update version in current project __init__.py
         package_path = package_name.split(".")
         init_file = Path(root, "src", *package_path, "__init__.py")
+
         update_version(version, init_file)
+        update_changelog_file(Path(root, "docs", "CHANGELOG.md"), version)
 
     return locals()
